Create script perturbation_1.py
This script is reads 'particle_input.dat' file and apply a perturbations in N_eu and Nbar_eu.
The present algorithm applies a positive random perturbation to particles moving in the x direction.
The same perturbation, but with a negative sign, is applied to particles moving in the oposite momentum direction.